moving things around a bit

+++ installers/go.sh
@@ -1,14 +1,15 @@
 #!/bin/sh
-curl -Lo go.tar.gz https://go.dev/dl/go1.26.0.linux-amd64.tar.gz
 
-sudo rm -rf /usr/local/go && sudo tar -C /usr/local -xzf go.tar.gz
-rm go.tar.gz
+if command -v pacman &>/dev/null; then
+  sudo pacman -S --noconfirm go
+else
+  curl -Lo go.tar.gz https://go.dev/dl/go1.26.0.linux-amd64.tar.gz
+  sudo rm -rf /usr/local/go && sudo tar -C /usr/local -xzf go.tar.gz
+  rm go.tar.gz
+fi
 
 # Install Air
-/usr/local/go/bin/go install github.com/air-verse/air@latest
+go install github.com/air-verse/air@latest
 
 # Install Wails
-/usr/local/go/bin/go install github.com/wailsapp/wails/v2/cmd/wails@latest
-
-# Install Wails v3
-/usr/local/go/bin/go install github.com/wailsapp/wails/v3/cmd/wails3@latest
+go install github.com/wailsapp/wails/v2/cmd/wails@latest
